pyproject.toml
[tool.poetry]
name = "mwdstdcore"
version = "0.5.0"
license = "AGPL-3.0-or-later"
description = "Core project for basic measurement-while-drilling corrections"
authors = ["MWD STD Inc. <info@mwdstd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Flask = "^2.1.2"
jsons = "^1.6.3"
attrs = "^21.4.0"
numpy = "^1.23.0"
scipy = "^1.8.1"

gunicorn = { version = "^19.9.0", optional = true }

[tool.poetry.dev-dependencies]
matplotlib = "^3.5.2"

[tool.poetry.extras]
gunicorn = ["gunicorn"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
